requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
langchain-community 
faiss-cpu
PyPDF2
langchain-google-genai